+++ src/index.ts
@@ -1,6 +1,12 @@
 export { ThemeProvider, ThemeConsumer } from "./context";
 export { default as useTheme } from "./useTheme";
 export { default as withStyles } from "./withStyles";
-export { default as makeUseStyles } from "./makeUseStyles";
+export { default as makeUseStyles, MakeUseStylesFn } from "./makeUseStyles";
 export { default as useColorTransition } from "./useColorTransition";
-export { maxWidth, maxHeight, minHeight, minWidth } from "./mediaQueryHelpers";
+export {
+  maxWidth,
+  maxHeight,
+  minHeight,
+  minWidth,
+  createDimensionQueryHelper
+} from "./mediaQueryHelpers";