+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1.3
+
+- Improve recreating styles
+- Added test
+- Added helper type `MakeUseStylesFn`
+
 # 1.1.1
 
 - Fix a `makeUseStyles` function type definition